C_V_4.1.7

+++ app/src/main/java/com/breezefsmshreebajrangsteeludyog/features/averageshop/business/InfoWizard.kt
@@ -394,7 +394,9 @@ class InfoWizard {
                             amount += list?.get(i)?.amount?.toDouble()!!
                     }
 
-                    val finalAmount = String.format("%.2f", amount.toFloat())
+                    //val finalAmount = String.format("%.2f", amount.toFloat())
+                    //mantis id 26274
+                    val finalAmount = String.format("%.2f", amount.toDouble())
                     finalAmount
                 } else
                     "0.00"
